Handle nil media list response
On a bad connection it is possible to get a nil media list response
which causes an unhandled error in the JSON deserialisation. Check
for string contents before making the call. Follows existing
convention of NSLogging an error and otherwise returning as normal.

It might be nice to retry here, but that would add complexity to the
sample for not a great deal of gain as it is not Cast-specific.
Catching the error prevents a crash though, which could confuse
developers otherwise.