Update cling / start looking at receiver name
WIP commit, updating the cling image and layout to more recent designs
and attempting to get the cast picking up the application receiver name.
Both these need more work - the cling needs layout fixes on the iphone
storyboard and testing and fixes on the ipad storyboard. The app name
currently isn't working at all as I suspect I'm misunderstanding the call
back, and the device manager needs to be an ivar to avoid it getting
reclaimed before its done executing.